requirements.txt
absl-py==2.2.2
alembic==1.15.2
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-JWT-Extended==4.7.1
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.9.2
libclang==18.1.1
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
protobuf==5.29.4
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.1.0
reedsolo==1.7.0
requests==2.32.3
rich==14.0.0
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
